trying to fix viz
