+++ real_spike/utils/__init__.py
@@ -1,2 +1,3 @@
 from .helper_functions import *
 from .latency import LatencyLogger
+from .socket import *
